feat(variables): improve variables selection string format & css
