+++ frontend/src/utils/map.ts
@@ -42,24 +42,6 @@ export function stepsColors(min: number, max: number, colors: string[]) {
   return returnVal;
 }
 
-export const listPossibleVariables = [
-  "bike_barsresaurants",
-  "bike_health",
-  "bike_posts",
-  "bike_schools",
-  "bike_transit",
-  "transit_barsresaurants",
-  "transit_health",
-  "transit_posts",
-  "transit_schools",
-  "transit_transit",
-  "walk_barsresaurants",
-  "walk_health",
-  "walk_posts",
-  "walk_schools",
-  "walk_transit",
-];
-
 export const geocoderAPI = {
   forwardGeocode: async (config: { query: string }) => {
     const features = [];
