+++ frontend/src/utils/variables.ts
@@ -1,0 +1,22 @@
+export const listPossibleVariables = [
+  "bike_barsresaurants",
+  "bike_health",
+  "bike_posts",
+  "bike_schools",
+  "bike_transit",
+  "transit_barsresaurants",
+  "transit_health",
+  "transit_posts",
+  "transit_schools",
+  "transit_transit",
+  "walk_barsresaurants",
+  "walk_health",
+  "walk_posts",
+  "walk_schools",
+  "walk_transit",
+];
+
+export function cleanVariableString(name: string) {
+  const str = name.replace(/(-|_|\.)/g, " ").trim();
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
